+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -31,7 +31,11 @@ function gwoCard() {
       "gwaio_start_hoarder",
     ];
 
-    const numCardsToOffer = 3;
+    const isLuckyCommander = function (inventory) {
+      return inventory.hasCard("gwaio_start_lucky");
+    };
+
+    const numCardsToOffer = isLuckyCommander(game.inventory()) ? 4 : 3;
 
     model.rerollTech = function () {
       var cardsOffered = 0;